Moved logs `controller` to the routes file
